Simplify typed destructuring
This commit updates the API of the `typechecking::destructuring` module
to return `UnifRecordRows` rather than a `UnifType`. We can do this
because currently only record can be destructured, and so the type
inferred from a destructuring pattern will always be a row type.

Additionally, this removes the need for `RecordTypes::Dyn`, so this
commit also removes that variant.