Fix typechecking and import resolution phase
The code handling typechecking and imports was a bit obscure: this
commit rework it, making clearer the assumptions and special cases.
Those two functions are special in that they have to be re-entrant: they
might be recursively called by themselves. Doing so, we fix a bug
(https://github.com/tweag/nickel/issues/1090) indicentally.

This commit adds a regression test for #1090, and make the test of
cyclic imports stronger (with several cyclic imports), which might make
the old version of the code break.