Simplify typed destructuring implementation (#1126)
* Only parse functions as FunPatterns when they destructure args

Previously any `fun` was parsed as a `FunPattern`, even if they didn't
do any destructuring. This commit changes the parser so that `fun`s
are parsed as `Fun`s unless they actually need to destructure their
arguments.

This means we no longer need to handle `Destruct::Empty` in the
typechecker (and additionally that it can likely be removed entirely
in the near future).

One additional note here is that while working on this commit, we
noticed a bug in the `FunPattern` branch of `walk`, where it was
assigning its argumetn the type of the function body. This had the
side effect of making a test in `typechecking.ncl` pass which should
otherwise have failed. As a result, this commit changes both the `walk`
implementation for `FunPattern`, as well as the test in question.

* Simplify typed destructuring

This commit updates the API of the `typechecking::destructuring` module
to return `UnifRecordRows` rather than a `UnifType`. We can do this
because currently only record can be destructured, and so the type
inferred from a destructuring pattern will always be a row type.

Additionally, this removes the need for `RecordTypes::Dyn`, so this
commit also removes that variant.

* Replace `Destruct::Empty` with `Option<Destruct>`

This commit removes the `Empty` variant from `Destruct` and instead
uses `Option<Destruct>` in its place. This makes it significantly
clearer where we expect to have a valid destructuring pattern.

* Replace Destruct::Record with RecordPattern struct

* Rename destruct module to destructuring

* Remove dead code

* Assign dyn types to pattern vars in walk mode