Add a dictionary_assume primop
A dictionary contract assertion gets rewritten to an application of
%dictionary_assume%. The latter synthesizes a magic open record contract
that has the same fields as the record on the left hand side of the
contract assertion. Each field is assigned the contract specified by the
dictionary. This magic record contract is then applied as usual to the
original record.