Fix query printing in the REPL
After the implementation of RFC005, the structure of metadata queries
have changed. Before, one could simply query an arbitrary value. Now
that metadata are attached to record fields only, the user must provide
a base record as well as a path pointing to a field. This commit fix the
previously broken metadata queries for the REPL by updating to this new
interface.