Fix dictionary contract semantics (#1141)
* Add a dictionary_assume primop

A dictionary contract assertion gets rewritten to an application of
%dictionary_assume%. The latter synthesizes a magic open record contract
that has the same fields as the record on the left hand side of the
contract assertion. Each field is assigned the contract specified by the
dictionary. This magic record contract is then applied as usual to the
original record.

* Add %go_dict% label primop

For better error reporting when dictionary contracts are violated, this
commit adds a new type path element variant `ty_path::Elem::Dict` to
denote moving to the field type described by the dictionary. It also
adds a `%go_dict%` primop for use by the `$dyn_record` in `stdlib/
internals.ncl`. This primop adds `ty_path::Elem::Dict` to the type path
in a label.

* Apply suggestions from code review

* Apply suggestions from code review