Remove useless code
The contract diagnostics are set by usercode during evaluation. However,
the LSP was branching of the presence of such diagnostics to enrich its
error message. It doesn't make sense, as the LSP doesn't evaluate any
Nickel code, so contract diagnostics are always empty.

This commit remove the corresponding useless if.