Use type annotations in record patterns during typechecking (#1176)
* Use type annotations in record patterns during typechecking

Previously these types were simply converted into contracts & then
applied at evaluation time. This commit uses the type annotations
at typechecking time to ensure that we correctly raise mismatch errors
if we encounter an invalid annotation while building the type of a
pattern.

* Fix typo in doc comment

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>

* Add tests for successfully annotated record patterns

* Highlight correct term on inconsistent pattern type

Previously if a pattern contained mutually incompatible type
annotations, e.g.:

```
let { x : { a : String } = { a : Number } } = { x = { a = 1 }} in a
```

we would highlight the term `a` in the error message. After this
commit we instead highlight the `{a : String}` term, which introduces
the invalid type annotation.

---------

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>