Follow bidir typechecking more closely
The current algorithm is a mix of bidirectional typechecking and
unification-base type inference. However, the relation to bidirectional
typechecking was loose at best.

This commit makes the typechecker follow more closely bidirectional
typechecking, and in particular the technique known as the Pfenning
recipe, such that the algorithm is simpler to understand and to extend.

Type instantiation is still quite ad-hoc, and cleaning it is left for
future work.