Implement array merging
In order to make merge idempotent, we've decided to implement a very
restricted form of merging on arrays as well. This commit is the first
part of the work, implementing point-wise merging for arrays (and
requiring their length to be the same).

In a second time, we plan to be even more restrictive, and refuse to
merge array elements that aren't the same (in particular, records should
have exactly the same set of fields). The logic and error reporting for
this task are dealyed to a follow-up PR.