Add Equal contract
Add an Equal contract to the stdlib, which checks that the value is
equal to some constant. This contracts try to preserve laziness, in that
checks are delayed for arrays and records in principle. Caution:
currently, the lazy check isn't implemented for record, so this contract
is actually just `fun c => from_predicate ((==) c)` when `c` is a
record, temporarily.