Implement the record_lazy_assume primop (#1211)
The contract.Equal requires a new primop, akin to the existing
array_lazy_assume, which adds a contract to the `pending_contracts` of
each field of a record. This primop also reverts recursive records, as
this new contract may alter fields, and thus trigger a recomputation of
the recursive record's fixpoint.

Doing so, we got rid of the `dictionary_assume` contract, which can be
subsumed (more easily and more efficiently!) by this new primop.