Add a merge label (#1217)
* Add `MergeLabel`

Add a datatype for merge labels, similar to contract labels, but
threaded through merges. It's not used anywhere yet.

* Adds a merge label parameter to the Merge primop

Adds a merge label, which has been defined just before, to the merge
operation. This is used to thread error reporting data throughout merges
(currently, only the span of the original merge).

This commit also makes use of this span for reporting incompatible merge
values.

* Improve merge error reporting

Add a distinction inside MergeLabel to indicate if a merge was
user-written (or descendent of user-written) or if it was generated by
the parser to combine a piecewise definition. In the latter case, the
error message is adapted.

* Disable clippy warnings about too many args