Put the stdlib under the std namespace (#1231)
* Introduce std namespace

This commits put the different modules of the stdlib, previously living
directly at the top-leve in the initial environment (`array`, `string`,
etc.) under the `std` top-level record. This makes the stdlib more
structured, give an unified entry point (including for completion and
querying) and avoid polluting the environment with name that could clash
with e.g. local function parameters.

This changes came with a number of trade-offs. The LSP isn't happy if
something like `std` isn't defined in one location, so we have to write
or at least generated actual code for `std` (instead of programmatically
building the record, for example).

We also want the users to be able to read this file, even if generated,
because otherwise diagnosing a contract error would be quite hard.

There were also other subtleties with respect to typechecking mutually
recursive modules. In the end, the stdlib is currently squashed into one
file (plus the module internals, but which is special) `std.ncl`, which
content is bound to `std` in the initial environment.

* Update the LSP to work with new std namespace

* Update examples to use the new std namespace

* Update benches to use std namespace

* Update integration tests for the new std namespace

* Update snapshot tests for new std namespace

* Last fixes for test after std namespace change

* Please the clippy God 🙏

* Update RELEASES.md - std namespace + RFC005

* Apply suggestions from code review

Co-authored-by: Viktor Kleen <viktor.kleen@tweag.io>

* Update stdlib/std.ncl

Co-authored-by: Viktor Kleen <viktor.kleen@tweag.io>

* Update stdlib/std.ncl

Co-authored-by: Viktor Kleen <viktor.kleen@tweag.io>

* Fix the LSP completion for `std`

* Please the Clippy overlord

---------

Co-authored-by: Viktor Kleen <viktor.kleen@tweag.io>