Separate between dictionary types and contracts
Since a recent addition, contracts derived from record types and
contracts derived from record contracts are different. The later are
local and don't propagate, while the former are propagated through
merging. This change calls for the same distinction for dictionaries.

This commit thus add two different dictionary syntax, the original
dictionary type `{_ : T}`, which derived contract is restored back to a
previous version where it's a mere map of the contract application of
`T` over the elements, and a new grammatical construct `{_ | T}` is
introduce, which contract is set to the contract used by `{_ : T}` in
the previous commit (which is a lazy, propagating contract).

As to make the impact of the change minimal, including for the LSP and
the error reporting of blame errors, `{_ | T}` is a variant of the
dictionary type which behave mostly exactly the same but for contract
generation. This way, the LSP, the typechecker and error reporting code
can stay mostly unchanged.