Convert existing ncl source files to annotated tests
This commit converts every existing .ncl file used in tests to an
annotated source file. This means we can change the glob we use to
pick up every .ncl file in the integration test subdirectory, which
makes test organisation significantly more flexible (and should
solve the problem of very long test file names by supporting arbitrary
nesting).

One additional change I had to make here, which increased the diff
significantly, is to add `repeat` functionality to the test annotations,
which allow a test to be re-run an arbitrary number of times. This is
to support tests where behaviour may be non-deterministic (e.g. it
occurs as the result of iterating through a `HashMap`).