move constraint checking code to parser
we check for constraints in contract checking and type checking. by
moving it to the parser we only have to do it in one place.

also, removed constraint checking for enum rows, which doesn't seem to
have been doing anything and also we don't want it.
`[| 'x ; e |]` does not preclude `'x` showing up in `e`

added in VarKindDiscriminant for those places where we do actually just
want the enum type. This is mostly for error reporting.