Add cached var levels data to GenericUnifType
This commit continues the implementation of variable levels. This adds
level-related metadata to unification types to avoid eager traversal to
update variable levels, following Didier Remy's "Efficient and
insightful generalization" algorithm.

This commit doesn't update the new variable levels of types correctly
yet, but merely add the infrastructure to store them and updates the
code to use the new types.