Fix LSP panic when importing JSON (#1382)
* Fix LSP panic when importing JSON

Nickel terms imported from a JSON file currently don't have any position
set, which was making the LSP panic. This commits stores the original
input format alongside cached terms, and makes this information
available to the public API. The LSP now uses it and abort any usage
analysis when the term comes from an external import.

* Add regression test for LSP external imports

This commit adds a test to the LSP test suite to check that importing an
external format doesn't make it crash as it used to.

* Set position when serializing from external format

* CachedTerm -> TermEntry

* Only make one hashmap access during linearization

Add an helper to `cache` to retrieve both the input format and a cached
term from the cache with only one hashmap access.

This commit also updates some related comments inside the code of the
LSP, which were not up-to-date anymore.