Inconditionally set usage for import in the LSP
A previous patch tried to avoid performing some code analysis on import
expressions of external format (JSON, YAML or TOML). However, it turns
out this introduced an off-by-one error in the LSP, which expects that
each and every node of the AST gives rise to an entry in the
linearization. This commits revert the previous patch and get rid of the
input format field in the cache, which isn't used anywhere anymore.