Break infinite loop in get_item
The get_item function from the LSP's linearization had a potential
infinite loop. This method used to take an additional maps from file ids
to linearization, and if the item wasn't found in the current
linearization, it would try to find the right one from the cache and
call the corresponding `get_item`.

However, when the item is nowhere to be found (which might have been
thought to be impossible, but it seems to be the case for external
import), the linearization found in the cache is exactly the same one,
and the function just do the same operation in a loop.

This commit:

1. Break get_item into two different method, a simple `get_item` which
   doesn't try to be smart and to look in the cache: just return the
   item from the current linearization, if any, and `get_item_with_reg`,
   which has the same behavior as the old one. However, instead of
   calling to `get_item_with_reg` when looking into the cache, it calls
   to `get_item`, which just stops here if there's no result. So, the
   loop isn't possible anymore.
2. Rename LinCache to LinRegistry, because it's not really a cache.
3. Took the occasion to remove some undeeded reference on the `Copy`
   type `ComplCtxt`