Add missing check for unifying equal type vars (#1400)
We didn't test that a record rows unification variable was equal to
itself before unification, which led to cycles in the unification table,
and thus infinite loop in the typechecker. This commits moved this check
directly at the level of `UnifTable`, so that higher-level functions
`unify_xxx` don't have to care anymore about it and are ensured not to
create such cycles.