In memory imports (#1374)
* Move the nls snapshot tests into the nls crate.

This makes cargo track the dependencies correct, and rebuild nls
before running the tests.

* Add an explicit "in-memory" tag in the file cache.

Entries loaded from in-memory sources will not be auto-refreshed
when an on-disk file changes.

This also changes the path normalization behavior, since the previous
implementation fails for non-existent files.

* Review comments

* Fix the tests.

Prior to this PR, the tests weren't getting run in CI,
and so the import_external_format test was broken.

This is a quick fix to that test, but:
- it seems pretty easy to make NLS panic on invalid input.
  This fixes a panic on invalid import paths, but probably
  invalid line/columns will still cause panics.
- this adds very basic support for server errors to the
  harness, but does not yet support tests that expect certain
  LSP errors.

* Add missing snapshot