SubstXXX -> one subst trait
Get rid of three variations on a `Subst` trait, to keep only one with a
template parameter. Introduce a substitution method to be implemented
which, in addition to substitution variables with something, also bubble
up new variable levels data, as they might have been updated by
substitution.

This commit doesn't actually implement the new substitution method yet,
leaving a trail of `todo!()` to be filled in follow-up commits.