Replace assert macros by functions
We defined various `assert_xxx` macro helpers in the early days for
tests. The only reason to use macros over simple functions was to report
the location of the caller located in the failing test, and not the code inside the helper, when
an assertion would fail. Since then we discovered the `#[track_caller]`
annotation which allows to do the same thing with bare functions. This
commit change such remaining macros to be bare functions.