[Refactor] Introduce RemoveRow trait (#1471)
* Put unify::(e,r)rows_add in dedicated trait

The free-standing functions `erows_add` and `rrows_add` are very
similar. The name and documentation was also misleading: what those
methods do is rather to remove a row than adding a new one.

This refactoring put them in their own trait, `RemoveRow`, and update
the documentation accordingly. Doing so, we also get rid of an optional
for the `UnsatConstr` error, which as unwrapped at some point anyway.

* Reword comment