Ignore missing field definition errors during documentation extraction (#1476)
* Ignore missing field definition errors during documentation extraction

My first implementation of a shallow evaluation mode for extracting
documentation tried to evaluate to WHNF every record field that
has a value. Then it reports any error that it encounters. But we
actually expect to hit `EvalError::MissingFieldDef` in normal partial
configurations because presumably an undefined field is intended to
be used recursively somewhere. With this change, we report all errors
except for `MissingFieldDef`.

* Add a comment describing the error recovery rationale