Actually persist output files for formatting (#1483)
* Actually persist output files for formatting

All the code for properly formatting files in-place was there, but
in the original PR I forgot to persist the output file. As a result,
formatting works for as a pipeline from stdin to stdout but when
an output file is specified, the formatted Nickel code is silently
discarded.

With this change we actually persist the formatted result.

* Add a rudimentary CLI integration test to make sure output files are created