Improve type error reporting
The implementation of the conversion from unification type to
human-readable types was a bit confused, using two name tables, not
updating correctly the set of already used variables, etc.

This commit simplifies the implementation, move free-standing functions
into `NameReg`, with the effect of correctly propagating position
information for user-written type variables and thus improving the error
reporting for the variable level mismath error.