Augment FieldInterface with more data
Augment the field interface type - for freeform argument passing in the
CLI - with additional metadata, such as being optional or defined, taken
directly from the original field's metadata. Add a somehow heuristic
(but simple) `is_input` method, which decides if a field is an input and
should have a dedicated argument.

Minor reworking/renaming of other part of the CLI export code around
freeform arguments.