Assume -> ApplyContract
At the very beginning of Nickel, the duality between a type annotation and
contract annotation was reflected in their original naming, the assume
construct (contract annotation) and the promise construct (a type
annotation).

This terminology isn't used anymore, and isn't very telling for new
contributors. This PR get rid of old references to `Assume` and
`Promise`, and rename the `Assume` primitive operation to the more
boring but more explicit `ApplyContract`.