Fix contract equality for unevaluated recursive records
The previous code failed when trying to check contracts for equality,
when those contracts had been reverted as part of merging (i.e. when
they recursively depend of other fields of the record), when they are
back to a function form.

To do so, we rely on the already existing behavior that contract
equality consider terms to be open, and that unbound variables are
tested for syntactic equality. Doing so, this commit also simplifies a
lot the previous machinery: we can apply the same idea to the initial
environment, and don't need to thread it through contract equality
checks anymore.