Add a `--field` argument to subcommands whenever it makes sense (#1712)
* Factor out extract_field() from query()

* Add selectable field to Program

* Add --field to supported commands

This commit finishes the series on adding `--field` argument to the CLI.
This argument is factorized as a new Customize implementer which
regroups commands that don't support general customization (overriding)
but still support selecting a specific field from the configuration.
It's added to the full customization as well.

In consequence, commands that didn't support selecting particular field
before now generically support the `--field` argument. The `--path`
argument for `--query` is removed and is just a `--field` argument for
`nickel query`.

* Fix not taking the --field into account when not customizing

* Formatting

* Fix tests for new --field argument

* Fix clippy warning

* Fix compilation errors, tests, and improve error messages

* Avoid duplicating 'field' declaration in clap