Stop using spot instances for release builds
We now had a failure to build release binaries due to a spot instance
getting canceled early for the second time. For now, I propose not using
a spot instance. While this will increase the cost, with it only getting
run once per release, I don't think there's cause for concern.

This change is change is already deployed to the AWS account to build
the 1.4 release artifacts.